# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "todolist-app API",
    "description": "개인 할일 관리 웹 애플리케이션 REST API\n\n## 인증 방식\n- **Access Token**: `Authorization: Bearer <token>` 헤더 (유효기간 15분)\n- **Refresh Token**: HTTP-only 쿠키 자동 전송 (유효기간 7일)\n\n## 할일 상태(status) 계산\nstatus는 DB에 저장하지 않고 서버에서 런타임 계산:\n- `NOT_STARTED`: 오늘 < start_date\n- `IN_PROGRESS`: start_date ≤ 오늘 ≤ due_date, 미완료\n- `OVERDUE`: 오늘 > due_date, 미완료\n- `COMPLETED_SUCCESS`: is_completed=true, is_success=true\n- `COMPLETED_FAILURE`: is_completed=true, is_success=false",
    "version": "1.0.0",
    "contact": {
      "name": "Yongwoo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "로컬 개발 서버"
    }
  ],
...
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Access Token 갱신",
        "description": "HTTP-only 쿠키의 Refresh Token으로 새 Access Token을 발급합니다.\n\nRefresh Token은 쿠키로 자동 전송됩니다.",
        "operationId": "refreshToken",
        "parameters": [
          {
            "in": "cookie",
            "name": "refreshToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HTTP-only 쿠키에 저장된 Refresh Token"
          }
        ],
        "responses": {
          "200": {
            "description": "토큰 갱신 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                },
                "example": {
                  "success": true,
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Refresh Token 만료 또는 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "message": "세션이 만료되었습니다. 다시 로그인해 주세요"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token (유효기간 15분, HS256 알고리즘)\n\n페이로드: `{ sub: user_id, email, iat, exp }`"
      }
    },
...
    "schemas": {
...
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "accessToken": {
            "type": "string",
            "description": "새로 발급된 JWT Access Token (유효기간 15분)"
          }
        },
        "required": ["success", "accessToken"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "오류 설명 메시지"
          }
        },
        "required": ["success", "message"]
      }
    },
    "responses": {
...
      "InternalServerError": {
        "description": "서버 내부 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "서버 오류가 발생했습니다"
            }
          }
        }
      }
    }
  }
}